SFB-88: Validations set in the editor are now also applied on the field form
