+++ app/components/field-validations-form.js
@@ -2,13 +2,11 @@ import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 import { restartableTask, timeout } from 'ember-concurrency';
 import { getFieldAndValidationTriples } from '../utils/get-field-and-validation-triples';
-import {
-  getTriplesWithNodeAsSubject,
-  getValidationSubjectsOnNode,
-} from '../utils/forking-store-helpers';
+import { getValidationSubjectsOnNode } from '../utils/forking-store-helpers';
 import { EXT, FORM } from '../utils/rdflib';
-import { BlankNode, Statement } from 'rdflib';
+import { Statement } from 'rdflib';
 import { ForkingStore } from '@lblod/ember-submission-form-fields';
+import { createBlankNodeForValidation } from '../utils/validation/create-blankNode-for-validation';
 
 export default class FieldValidationsFormComponent extends Component {
   @tracked store;
@@ -54,7 +52,6 @@ export default class FieldValidationsFormComponent extends Component {
     const fieldTtl = this.store.serializeDataMergedGraph(
       this.graphs.sourceGraph
     );
-    console.log(`field ttl`, fieldTtl);
     const applyStore = new ForkingStore();
     applyStore.parse(fieldTtl, this.graphs.sourceGraph, 'text/turtle');
 
@@ -71,17 +68,19 @@ export default class FieldValidationsFormComponent extends Component {
 
     const validationsToApply = [];
     for (const validation of validations) {
-      const validationNode = this.createBlankNodeForValidation(
+      const validationNode = createBlankNodeForValidation(
         validation,
-        applyStore
+        applyStore,
+        this.graphs.sourceGraph
       );
+
       const statement = new Statement(
         this.fieldSubject,
         FORM('validations'),
-        validationNode,
+        validationNode.node,
         this.graphs.sourceGraph
       );
-      validationsToApply.push(statement);
+      validationsToApply.push(...[statement, ...validationNode.statements]);
     }
 
     const validationsToRemove = [];
@@ -110,28 +109,5 @@ export default class FieldValidationsFormComponent extends Component {
       applyStore,
       this.graphs.sourceGraph
     );
-  }
-
-  createBlankNodeForValidation(validationNode, store) {
-    const blankNode = new BlankNode();
-    const triples = getTriplesWithNodeAsSubject(
-      validationNode,
-      store,
-      this.graphs.sourceGraph
-    );
-
-    const blankNodeStatements = [];
-    for (const triple of triples) {
-      const statement = new Statement(
-        blankNode,
-        triple.predicate,
-        triple.object,
-        this.graphs.sourceGraph
-      );
-      blankNodeStatements.push(statement);
-    }
-    store.addAll(blankNodeStatements);
-
-    return blankNode;
   }
 }
